feat(graph): expose functions to render pdv & error page
